Add a catch all cases super function
After discussions on the API.

Adding a function:
   add_predictor_constr(model, predictor, x, y=None)

This would take any predictor that we understand as argument and add
the correponding constraints and variables to the Gurobi model.

It returns the object that contains the descriptions of the additions to
the model (the type of the object depends on the type of predictor).

Note that the output variables is now an optional argument. They will be
created by the modeling object if they are not passed. They can then
be retrieved with: predictor_constr.output