Accept 2-D MLinExpr as input for predictor constr
This is to avoid having to create variables for fixed features.
Thanks Simon for the example on how to build it.

At this point this is just to explore and play with the idea.

Had to fix standard scaler. Note that this is still a bit dumb because
scaling it still creates variable to get the scaled version of a
constant.

I think that the decision tree stuff shouldn't work and also polynomial
features.