Add an object that is the addition of a constant ndarray and an MVar
To get rid of the runtime warning about using linexpr in a slow way.
Created this object that supports only the few matrix operations that
we need so that they can be made faster.

It also has the advantage that it is easy to do a better
implementation of decision tree where we don't add indicators with a
fixed consequence term.