some convolutional net
Started doing a convolutional net with keras.

notebooks/adversarial/adversarial_cnn.ipynb

contains a working example that we would want to model with gurobi_ml

A conv2d layer is in progress (doesn't crash and does something).

The big mess now is variable dimensions. Gurobi ML was more or less
designed around the idea that the input of the ML model is 2-d
(n_examples, n_features) but of course convolutional layers are 4-d
(n_examples, height, width, n_channels)
The 2d assumption is all other the place...

Missing layers:
- MaxPooling
- Dropout
- Flatten

Also for the conv2d only did with valid padding and the for loop is